[postgres] Combine `scanner.go` with `cast.go`
